config(prr): remove
I did not like that it did not have proper syntax highlighting. It was also a strange interface to require separate files. I think I can do better by trying to create a script that will check out a pr, squash all the commits to the base branch it is merging into, then resetting to one commit before. This will make it so the git dir has all the changes as unstaged and I can review all of it as if I made the changes and about to push them.