fix(prettier): properly set fallback
There was misunderstanding with code where I thought that I was setting
fallbacks to literal format commands. I was actually setting fallbacks
to format configurations. There is no such `/nix/store/.../bin/prettier`
config. That is not a valid formatter config.

It is in the formatter config that you should set fallback binaries.