fix(lsp.attach): prevent initial attachment
By attempting to do things the "right way" of not monkey-patching
functions, we run into issue that the LSP attaches before our code gets
a chance to stop and detach it...

If we instead forcefully override the attachment behavior, we can get
desired outcome.

I am not sure consequences of this.

Configuring every lsp to not intially attach is simply not a viable
solution. We really need a way to interrupt attachment.